Backs out HAVE_LD_EH_FRAME_HDR recent commit
FreeBSD has a function dl_iterate_phdr in libc that DragonFly lacks.
The previous state of the repository couldn't be built and this missing
function was in part why.  I am going to leave libiberty.h modification
intact, but there were some trailing spaces after the forward slash that
gcc was complaining about, so I fixed that as well.