Returning to use of GCC's stddef.h to replace Dfly version
The t-makefile fragment for Dragonfly was configured to let GCC use
DragonFly's stddef.h header instead of the one provided by GCC.  This is
what NetBSD sets in the pkgsrc version of GCC.  Dragonfly developers got
around it by adding the stddef.h file back via the include-fixed
directory which is not encouraged.

By having neither the include-fixed version nor the GCC version, it was
not possible to build C++ library due to a missing function.  I decided
to finally fix the ginclude/stddef.h to work for DragonFly.  The changes
that were done previously weren't correct and prevented wchar_t from
being properly defined, plus it had an unnecessary <machine/stdint.h>
include.  I modified it to work like FreeBSD 5+ and removed the override
in the fragment file.  It's currently a placeholder.