Updated ada/init.c to avoid compiler warnings
Once GCC 4.5.0 bootstrapped, the compilation stopped because the compiler was
treating errors as warnings.  The problem with init.c was twofold.  One
function had an unused argument that wasn't annotates as unused (it also had
another argument that was).  The second problem was the "old-style" function
declaration.  Making these small changes allowed the compile to continue. I
suspect Adacore will make a similar change soon.