Build scripts updated to 1 APR 2010 snapshot
The 1 APR 2010 features a lot of fixes as compared to the previous snaphot.
In fact, this is what the GCC 4.5 release candidate is based on.  Due to a
C++ regression, I'm declaring that GNAT AUX needs this snapshot as a minimum,
so the sample build scripts have been updated to reflect that.