Removing explicit thread specification in FreeBSD
For some reason, FreeBSD AMD64 complains when thread model single is
specified.  Since for i386, this the is thread model that the configuration
falls back to as a default, I decided to delete the enable-threads item
altogether and just let gcc pick the default for both i386 and AMD64 for
FreeBSD.  I also removed the --disable-bootstrap configuration item that
really wasn't supposed to be in there to begin with.