Fixing Xu's frame unwind patch
While running the new patch through the paces with FreeBSD AMD64, the build
broke at the new patch.  It seems 3 of the methods were attached to the wrong
structure, which doesn't bode well for the quality of the patch.  It could
not have passed compilation unless the _Unwind_FrameState structure was
different when the patch was built.  This updates the last commit.