Added AMD64 targets for FreeBSD and Dragonfly to gccint Makefile
I thought the current FreeBSD patches were already in the 4.5 series of GCC,
but after closer inspection, I found that the x86_64 target for FreeBSD
actually was not part of the gcc-interface makefile.  The
system-freebsd-x86_64.ads file that it references was though. Even though we
don't yet have support for DragonFly x86_64, I added the target now.  Likely
the system file for DragonFly x86_64 will be similar if not identical to
FreeBSD.  The missing target was discovered during ACATS testing of AMD64
build.