Updated thread interrupt priority defs for FreeBSD AMD64
Currently FreeBSD x86_64 is failing around 25 tasking tests.  The one
file that is different between i386 and x86_64 is the system-*.abs
file.  After a lot of searching, I found the thread priority definitions,
including the thread interrupt range, in <sys/priority.h> on FreeBSD 8.0.
This commit updates the system-specific information.  I've no idea if this
will solve the task segfaults or not.