Updating FreeBSD i386 Interrupt range
The interrupt range for FreeBSD is not architecture-specific.  i386 also has
a maximum interrupt priority of 63.  This likely changed between FreeBSD 4.x
and 5.x, but nobody ever updated the system file.  Due to the stupid style
rules, I had update the AMD64 version because the lines were too long.  I also
updated the Backend_Overflow_Checks to True since the original patch was made
before the i386 system file was updated with regards to this parameter.