Fixing minor compiler complaints on Ada files
The compiler was complaining that in some cases a variable or more could be
used before it was initialized with a value, but in reality this wasn't the
case due to the intended usage of the procedure.  However, to avoid warnings
from popping up, the variables were initialized or otherwised guaranteed to
be established with this change.