Cleans up adaint.* again
This adaint.c file must not have been maintained that well because it's still
emitting warnings.  This commit adds a function to the header file where it
was never defined (I think this means it was never used), and also changed
a string from static to const to avoid the "Initialization discards qualifiers
from pointer target type" message.