Cleaned up DragonFly GCC config; eliminated 64 file
The standard Dragonfly header for the i386 configuration had 64-bit switches
embedded in it, as well as a small 64-bit configuration file.  I didn't see
the sense in having both, so I deleted the 64-bit file and moved the
differences to the main header file.  I updated the config.gcc to match as
well as making a small cleanup there as well.  FreeBSD suffers from the same,
but I'll leaving it alone for now.