Added x86_64 system file for DragonFly BSD
In an earlier commit, I added the system-specific file list for the 64-bit
DragonFly BSD build, but it referenced a file that did not exist until now.
The 64-bit version of the system file only differs by the size of the word,
the exact same difference that FreeBSD is showing currently.