Adding sample solaris build script
Unfortunately building GNAT AUX on Opensolaris SXCE 130 is producing a compiler
that is DOA with regards to making executable files.  Perhaps it would work
on OpenSolaris 2009.06.  The reason for the problem isn't known, probably a
simple configuration error inside gcc.  However, this script still may be
useful in the future, so I'm adding it.  It's just a sample script anyway.