Target and architecture aren't necessarily the same
When I combined the dragonfly.h and dragonfly64.h files, I got mixed with with
the TARGET_64BIT macro.  Normally a 64-bit target is compiled on a 64-bit
build machine, so these values are normally the same.  However, since it's
possible to compile for a 64-bit target on a 32-bit machine, the __x86_64__
macro needs to distinguish the differences between the two files, not the
TARGET_64BITS macro.