GCC repository follow to-do list
Eric is getting ahead of me and I'm kind of stuck because I think his changes
would break GNAT AUX if it's overlayed over the latest GCC 4.5.  I'm going to
have to think about the update strategy.  Likely I'll tag a good spot for use
with 4.5 and then catch back up with the trunk.  The alternative is to carry
a bunch of non-gnat c files and I don't want to do that if I can help it.