Draco Driver work, new system specific file
The biggest add here is a new package that is system specific.  The idea
is that the ads file will actually be a template and the configure script
will populate it with system specific information.  This will be used in
place of the horrible C preprocessor macros strewn all over gcc.