Custom pipe command for Draco Driver
To support the -pipe command, the Draco driver needs to be able to pipe
(obviously).  However, this doesn't appear to be be built into the GNAT
library, although its components were.  This pipemechanism package that
I put together will take arguments to spawn two programs with one's
output being connected to the input stream of the other.  It should even
work for Windows which the GCC driver does not support.  However, I don't
have GNAT on windows readily available, so this isn't tested yet.