Driver snapshot
I got the CC1_SPEC implemented for alpha, arm, i386, ia64, and mips.  The
only backends missing are powerpc and sparc, but basically I don't need to
spend any more time on it because this all will be removed once GiGi is
ripped out and replaced with DLC (I think).  In the same vein, I'm not going
to implement ASM_SPEC or ASM_FINAL_SPEC (which applies only to alpha/osf5
anyway) because it's not a permanent contribution.