config.gcc update along with i386.c
Two files which are constantly being updated are config.gcc and i386.c.
We need to track the latter to keep the Dragonegg patches in place.
Hopefully this requirement will go away in the future, but it is
still necessary to use DE for now.

The i386.c changes are updates over the last month, about 9 commits
over the last 9 days.