Modified copyright ownership of changes
Since "AuroraUX" is not officially (or legally) an organization, I changed the
copyright assignment to myself.  I left the url "www.auroraux.com" to distinguish
me from the other John Marinos of the world and to also provide a useful link
about who to contact about questions with the updates.