Added help, verbosity, and assembler argument to Draco Driver
I added the help screen to Draco.  In order to make the style rules stop
complaining, I had to wrap the lines and now the routine is ugly.  I
moved the verbosity header out of the process files section so that it
would work even without any file names like gcc does.  I also added the
-Xassembler option which was missing from the switchmap for some reason.
It pretty muchs works like Wa, (almost the same function) but it only
allows single arguments.

I also added the "configured as" for the verbosity history.  The CMAKE
script will have to pass this along.