Random accumulated changes, part two
Since I just committed 46 files that have nothing to do with each other,
I figured I could do that again to finish off this huge commit session.
This is the remaining set of files that have changed over the last two
weeks excluding a couple of files that I want to commit separately.