Reverted problematic ACATS exec changes
In the recent update to the run_acats.sh script, a replacement "which" utility
was provided.  It caused an error for FreeBSD 8, causing the ACATS not to
execute at all due to failure building the support files.  For now I'm going
to assume all platforms have a functional "which" utility and just delete
this replacement function.  It fixed the FreeBSD issue.