options.c/options.h autocreated
There's a strange issue where if I attempt to create a static library from inside a cmake function
it fails because the macros don't get sent in.  It's probably some kind of "scope" issue but it's
very weird.

Regarding the title, the libgcc_eh.a library requires options.h and that's one of the autogenerated
files.  I limited the inputs currently to ada, common, and cpu options.  That means link time
optimizers and the c-family options were omitted.  I've got an issue with auto-host.h, which is
required by tconfig.h (another autogenerated) for some reason I don't know yet, and it's generated
by autotools.  Hopefully I can resolve this issue easily.  I'm a bit nervous about it.