Commit represents a huge frustrating day.
After some initial progress, CMAKE just wasn't working out.  It turns out
that it works a bit differently that autotools where there really aren't
any "targets" that you can request after the configuration phase.  For
example. on gcc you could configure and later type "make all" or
"make bootstrap-lean" or "make cross-tools".  Not in CMAKE.  There
are targets, but they are for unix commands only.

Not only that, I was having issues caused by thinking the script was
sequential.  No, it's really based on directories.  You can add
definitions and flags, but on a per-directory basis.  If you add and subtract
the system acts like you never added it at all.  That was a bitch to figure
out.  I had to completely redesign the script to have one cmake script per
directory.  I'm still feeling my way through this.  With the creation of
libgcc.a and libgcc_eh.a, it seems to be working as expected, finally, so
maybe I am progressing again.