Implemented dependencies and crt objects
As the title says, the big addition was the compilation of the crt c-object
files.  However, I also implemented stage-based dependencies, all which
trigger under the "ALL" target named compiler.

I just realized that making a cross or xeno compiler is going to require
an external cross compiler for the c-objects.  I didn't consider this and
it obviously makes creating a xeno-compiler much, much harder.  However,
the plan is just to provide the full path to the special compiler as a
requirement.  I think maybe that cross-compilers aren't affected by this
and we can continue to use the system compiler with a foreign as/linker.
Let's cross our fingers.