Added generation of DRACO specification file
The main focus of this commit was to have the DRACO system configuration file
be generated during each stage build.  It's an Ada file and this is working
nicely.  The next step is to have the driver get generated to see how CMAKE
handles building normal Ada files with no C-objects to be linked.