Added gnatlink, gnatbind, and extra ada compiler options to script
Within the driver script fragment, the gnat tools (link and bind)
are located and stored in a script, as well as the Ada compiler
itself.  The tools and compiler paths can be separately specified
otherwise the current search path is use.  If one of them is not
found then the script throws a fatal error.