Good Vibrations.  Cmake successfully builds Ada application
Despite CMake only knowing C,C++, and some fortran, it's possible to use to
to build Ada programs without using gnatmake.  This commit adds the new
Ada-based driver to the script and while there were some tricks, it worked
just fine.  The new driver is built in its proper sys root, which was
possible due to the use of the custom command.