GNAT1BIND dependency generator
I am taking a different approach to building gnat1 and gnatbind than is done
for gcc (I hope it works!).  I provide the script with the files to compile,
in order.  I also provide the script with the object file dependencies so
that CMAKE knows that the custom command must be executed and also the
composition of the executables.

This information is generated using a PHP script that I wrote.  Inputs are
the ada object definitions from the gcc Make-lang.in file, split into
separate text files.  The idea is that you can copy and paste from the gcc
script.

I still need to add another input stating if the file is to be generate
so that the location of "L" and "R" will read "L" for local in those
cases.  Still a work in progress, but nice.