Save progress - needs overhaul
I need to rework this script a bit at a base level.  It's not going to work
for cross-compiler as it is due to the c-compiler being hard locked with
the standard add_executable, add_library functions.  I need to make my own
functions and just ignore these.  Also, need to rearrange build order,
move pregens that are not specific to HOST to be built only once, etc.