Added i386/freebsd autohost constants
These autohost files are specific to GCC.  GCC generates each one, testing
a couple of hundred machine-specific parameters.  I just paste the results
of a modified config.in file as a temporary measure.  When we switch to LLVM,
these targets will go away (llvm takes care of it, draco should be a lot
more insensitive to the targets)