Ending this cmake effort after 2 more libraries added
A few days ago, I added two more libraries to the CMake script, but after a
failure on FreeBSD that didn't exist on Solaris, I realized that each library
needs it's on target-specific config.h file which is built by autotools.  It
is just too much of a pain to recreate this or copy over the config.h file
like I did with the main autohost.h file.

I've proven CMake will do the job just fine, so I'm going to stop work on this
phase of Draco.  Once GiGi is completely replaced, I'll pick it up again,
eliminating all the GCC backend stuff (which is 80% of the script probably).