Binding to LLVM core library (C only)
Rather than reuse the previous binding attempts (swig in llvm and start of
Rothwell's work), I redid the C binding for LLVM from scratch using the
inbuilt GNAT C/C++ binding generator.  Unfortunately only C is an option
right now because the C++ binding generator doesn't know how to handle all
the variations of the C++ template (specifically on the Module.cpp) file,
plus it's pulling in all sorts of target machine headers.

With the C binding, there was only one type that was being pulled in from
the target machine (unsigned byte) so I just hard-coded that to make it
machine independent.  The C API seems pretty tedious compared to the C++
API (not to mention virtually undocumented) but at least the interface is
limited to a single specification file.