Testing LLVM C API
I was trying to recreate the LLVM language tutorial for a few days and I had
some success, but in the end I was spending too much time messing around with
code that wasn't important (but it was good practice).  Anyway, I decided to
interface with the LLVM libraries directly using the tutorial as a model to
see if I could get the same output, and so far it's working.  I'll use this
as a reference for future work.