First attempt at AST converter stub
I'm slowly understanding how GiGi works and starting to form an idea on how to
replace it.  At this time, I'm planning to basically follow the same procedure
that Gigi does (C), but do the analogous procedure in Ada by interfacing with
the LLVM libraries.  Unfortunately I still have to learn a lot about both LLVM
and GCC Gimple in order to accomplish this.  However, to show some progres,
here is the DLC (Draco-LLVM (AST) Converter) specification.