Globals Scheme + TreeSync Addressing
This commit represents a lot of trial and error, and a lot of time.  The
key was the attributed 'Unchecked_Access.  With this, I was able to create
TreeSync locally, but refer to it globally.

Obviously this is an unsafe situation because techically the memory for
TreeSync could get claimed while the dglobal pointer still points to the
memory.  In practical terms it is not an issue because all the subroutines
that reference the TreeSync pointer will do so while TreeSync is still in
scope.  So it's unsafe, but the potential dangle will not occur.