Introduction of TTree structure
I was spinning my wheels until I understood the gcc tree variable is a
pointer to a union structure.  Nothing seems to fit perfectly, so I am going
to attempt to mimic this approach with the new TTree structure.  So far, so
good, but I haven't scratched the surface yet.