Snapshot - introducing module, phase out padding?
Trying to mirror how GiGi works with gcc might be impossible.  The gcc tree
structure holds a ton of information, e.g. padding and the original values.
I don't know if this concept even exists in LLVM, or if it does, it seems
to be hidden (nicely) inside the library.  With the hope that tracking
padding and a myriad of other attributes is simply for the benefit of gcc and
doesn't apply at all to the LLVM Intermediate Representation, I will attempt
to phase this out, which will probably require backing out some previous
work.

On a lighter note, I create an LLVM module in the DLC call procedure.  Even
something as simple as this required a couple of hours of study, but at least
I'm learning the GNAT internals bit by bit.