PHP utility to condense test "summary" logs
We don't care about listing all the passed test for ACATS and gnat.dg, we
only need to know about the ones that failed or were unsupported.  This little
php utility quickly concatenates filtered versions of these reports, so we'll
use those instead.  As an added bonus, we get rid of all the snapshot
subdirectories too.