Moving from strlcpy & friends to strncpy
I just learned about what an asshole Ulrich Drepper is, and why glibc doesn't
host these functions that are in all BSDs and Solaris.  I still want GNAT AUX
to be compatible with Linux, so I'm dropping down to the inferior version that
is on Linux.  Seeing how we used strcpy before, we should be fine.