Safer C routines from OpenBSD for GNAT
These changes stem from changes to adaint.c, which mitigate the risk of
buffer overflow.  It's more work to carry them forward as a coupld of files
weren't being tracked/modified, but the changes should be universal for all
the platforms I'm trying to support.  Maybe less portable (unlikely) but I
don't have to worry about the fringe cases for a while.