Oct 30 Snapshot test results (some)
Well, the FreeBSD64 server is still clicking off with an external powered
hard drive, so it's looking more and more like a failed CPU fan is the
culprit.  I haven't been able to compile GNAT without it unpowering a few
few times.  I did get results for DragonFly 32, 64 and Opensolaris 32.