Improved NetBSD Support, SSE Support for DBSD
Nearly two years ago, a guy named Aran Clauson seemed to get NetBSD running
GNAT and provided almost working patches to FSF.  Of course, they blew it off
but he did a lot of work with the tasking.  This also gives me a starting
point to try to fix FreeBSD64 tasking as well.  A cleaned up version of his
patches are introduced here.

I also am testing SSE support on DragonFly BSD (Should have been a separate
commit, oh well, it was in the same Makefile).  If this works, expect it to
spread to all the BSDs.