Good news, OpenBSD64 tasking error resolved
My suspicion about the OpenBSD tasking error was confirmed; it was due to a
mismatch on the datatype of the time_t type.  OpenBSD defined this as an
integer rather than a long integer and it caused a negative time number
resulting in an time either invalid or impossibly far off, making the system
hang.  Creating a BSD specific os-primitive probably was the main fix.

Unfortunately it's duplicated in two places, not really a good design.
Ideally you should only have to set it in one.