Reversed Ada 2005 Issue 157
This binding interpretation changed a previous warning into a fatal error,
making it impossible to build gprbuild under normal circumstances.  I've
decided to permanently patch all GNAT ports with a reversal of this change
until a new gprbuild that is buildable by GNAT FSF is published, which will
probably be next summer.