Final rebuild of Dragonfly32, results
I rebuilt DragonFly32 on DragonFlyBSD 2.8, without the AI-157 patch as I
discovered this was unneeded.  I now use "aux" for all vender strings of the
triplet to prevent a real gcc from overwriting it.  Also, it now passes the
sse tests (I patched the bug in the testsuite).

Still perfect!