FreeBSD Task priorities have been incorrect.
I'm still not completely certain, but it looks like I confused the scheduling
interrupt priority with real time priority.  The former ranges 0 - 64 and
the later ranges 0 - 31.  The former was probably causing the priority
failures on freebsd32.  Reverting to 0-31 resulted in a clean testsuite, but
apparently this is one of those tests that pass sometimes and sometimes it
doesn't.