Perhaps unnecessary tweak to xmlada package...
This package built okay, but I removed the wrapper binaries (removing the
complete wrapper caused an error due to missing untransform.sed file).  The
result didn't change I don't think, but at least it compiles a lot faster.
This will probably be the standard for all of my packages.  I'll try
backporting this to the gnat-aux port and see if that fixes netbsd32.