Modified AWS IPv6 option name
The AWS IPv6 was getting enabled by default even though the option wasn't selected.
This was probably because a standard option was used (inet6).  I thought it would
have had to have been specifically requested, but apparently it is a system wide
setting so I just change the name to make it work as expected.