Fixed problem with gnat-aux pkgsrc makefile
The iconv package has a builtin version that isn't compatible with the
gnat-aux bootstrap compiler.  This tweak forces pkgsrc to install libiconv
from source and hopefully fixes the NetBSD32 gnat build problem.

To be fair, the original problem had a slightly different signature, but it is
possible that the two problems were related.  I'm testing on a new NetBSD
installation right now.