Finally, fix for 32-bit platforms
After an exchange with Eric Botcazou, he told me tracebak.c was highly platform
specific and to take a look at it.  Recently Linux32 had been switched over to
the GCC unwinder and the AMD64 targets were all using it as well.  So rather
than try to tailor the generic unwinder, I added a specific callout for the
four BSDs and the initial results on AWS look good.

Also I flushed out a minor issue with the BSD socket simplification.