FreeBSD64 GNAT AUX Fixed!
With the help of the threading guys on FreeBSD, we determined that the
problem was trying to execute tasks from the stack which is not allowed in
FreeBSD (but not respected in i386).  Armed with that information, I was
able to make a macro to make part of the stack executable on the fly like
NetBSD and OpenBSD, and now FreeBSD passes all tests!  Proof to follow.