Updated FreeBSD GNAT AUX port with new bootstrap compilers
I built two new bootstrap compilers for FreeBSD32 and FreeBSD64.
They have the advantage of being statically linked except for libc.
That means the bootstrap compiler should not break as mpc, mpfr, gmp, z,
and iconv get updates over time.  If FreeBSD 9 upgrades the libc, then
we might have to add a compat8x for FreeBSD 9 later.  This is easy to do.