Build libgcc_s.so on DragonFly BSD
It was tricky to build libgcc_s on dragonfly due to the fact this file is not
present on the system (only a probably for disabled bootstraps).  But passing
a value for LD_LIBRARY_PATH that points back to the build/gcc directory, it
enabled the complete build of libgcc_s and the package list was adjusted
accordingly.

This commit also fixes a previous error with the t-netbsd file, and restores
the top part to how it currently exists in the gcc codebase.  My previous
change not only was unnecessary, it was actually wrong.  :(