Added NetBSD AMD64 bootstrap
The NetBSD AMD64 bootstrap was created and tested, and the changes reflected
here.  I forgot to remove the LIBGCC_SPEC from dragonFly, so I'll have to
update the source again.  I will probably roll to Jan 20, even though that sort
of invalidates the NetBSD build tests (although I will test that on DFly)