Prepare BSD for using dl_iterate_phdr capability
The worst part about this patch is that it requires a change to
gcc/configure.  Rather than carry another file that basically changes
weekly, I'm going to carry it as a patch and apply it when the
source is getting built.

I've created a new t-fragment called t-exceptions and now I can
ignore t-netbsd and t-freebsd (and don't need to create d-dragonfly).