Disable dl_iterate_phdr for NetBSD and DragonFlyBSD
It turns out that all the shared libraries need to be built by a compiler
that is already dl-iterate-phdr enabled (e.g. the system compiler).
This is basically only a requirement for gnat as c++, etc., have tricks such
that result in a working compiler and don't require -static-libgcc for
everything except shared libraries.

Once DragonFlyBSD has dl-iterate-phdr in the system compiler (probably with
4.4.6) then this will be re-enabled.  The same for NetBSD.

As a side commit, the stage1 flag -fkeep-inline-functions was removed as it
was preventing a full bootstrap on several archs/platforms.  It's purpose
was for helping non-gcc compilers bootstrap, so it didn't even apply.