Draft pkgsrc update to gnat-aux
This was the intended final version of gnat-aux pkg after the decision to
disable dl_iterate_phdr was made.  However, it's a good idea to implement
the conditional dependancy on dejagnu like we did for FreeBSD, so I want to
do that before we test.  This is just an intermediate commit then.