Define tmp directory for Android
Android does not use /tmp as the default temp directory.  In fact, it doesn't
define a temporary directory at all.  This patch will try first the
equivalent of /data/local/tmp which is a common folder location that requires
root access to use normally.  If that fails, it will try external storage
location (e.g. /sdcard/), but this will require the binary has write
permissions, which can be obtained by the application wrapper of
the nativeactivity manifests.