Eliminate NetBSD-specific taprop configuration
NetBSD and gnu/kFreeBSD share the situation where the
get/set protocol and get/set ceilingpriority for threads
are not defined.  These are optional in the posix standard,
so they just weren't implemented on these systems.

This was previously taken care of in NetBSD by creating a
variation of the s-taprop-posix.adb file for netbsd that didn't
reference these functions.  Gnu/kFreeBSD took a different approach and
created dummy functions that always return zero (success) but don't
modified the thread at all.  Logically, these two approaches are
equivalent, but L.Brenta's method has the advantage not having to
track future changes of s-taprop-posix.adb and also being able to
accommodate future implementations of these functions a little bit
cleaner.  It's also probably easier for future maintainers to
realize what was done here.

Note that getprotocol and getprioceiling is not used anywhere in
the gnat codebase, so these were removed entirely from the
s-ointe-netbsd?.ads files.  The other BSDs still define them
unnecessarily for completeness.