Reinventing the wheel
I had forgotten that I already created a GPS assembly script,
so I ended up creating it again.  There's also a new approach
for using gps_gpl which already comes with gnat sources, and
70M of documentation.

The new script limits the doc to pdf only, and repackings the
tarball.  Also included are ways of getting the head versions
of gnatpython, xmlada, etc.