Add -R support to project files
The XML/Ada libraries proved that gnat is incapable of supported DESTDIR,
and previously it would not acknowledge the use of the -R switch, which
is what gnatlink uses to prevent rpath from being built into a shared
library.

This change here will build in standard adalib path into the rpath, but
it will not add in implemented library paths if the -R option is
detected within Library_Options of a gpr file.