Post-release bug fix + resolve ambigious function
The gcc-interface/Makefile.in shared gnatlib target was released with
a bug, so this is the patch.

Also, there are near-dumplicate clock_gettime functions and they clash
so the rough workaround is to give one a new name.  DragonFly uses a
different datatype than the other BSDs (unsigned long vs int) but
it should be compatible.  The a-exetim version is hardcoded as C.int
rather than clockid_t due to the use of a constant.