gcc47: Create unique "Ada" driver
Using the default "gcc" driver for the various Ada programs such as
gnatchop, gnatlink, etc. isn't working well.  They catch the system's
compiler first.

The solution proposed here is to create a hardlink of the gcc driver
called "ada" which is obviously not a standard driver name.  Then the
various gnat programs will use this driver instead of "gcc".

A new gprbuild will have to be made to support it -- current ones look
for drivers called "gnatgcc".