Add "static" build option to lang/gcc-aux
There was something wrong with exception handling on NetBSD, seems to be
a problem with the platform itself.  The solution was to add the ability
to build gcc-aux statically so that dl_iterate_phdr error handling
works.

It's a generic options that it should work on DragonFly as well if
desired.