Remove dummy convert_addresses function from adaint.c
With the new traceback_symbolic.c file, there was a duplicate
definition of convert_addresses function.  At the very least, all the
system checks after IS_CROSS would have to be deleted, but I think
keeping it in traceback_symbolic.c is cleaner, so remove it from
adaint.c file.