Adjust FreeBSD unwind so it works on 32 over amd64
The KERN_PS_STRINGS method does not work when the i386 compiler is
running on amd64.  This is because the signal trampoline is in a
different place from the "pure" i386 system in a shared page.  Use
the frame pattern method to achieve a perfect ACATS/gnat.dg score
on that setup.